All fixed

+++ src/Movement.java
@@ -1,16 +1,16 @@
 
 public class Movement {
 	Coordinates from, to;
-	double quantity;
+	long quantity;
 	
-	public Movement(Coordinates from, Coordinates to, double quantity) {
+	public Movement(Coordinates from, Coordinates to, long quantity) {
 		this.from = from;
 		this.to = to;
 		this.quantity = quantity;
 	}
 	
 	public Movement(Coordinates from, Coordinates to) {
-		this(from, to, 0.0);
+		this(from, to, 0);
 	}
 	
 	@Override
